+++ app/src/main/java/com/example/tsaving/LoginActivity.kt
@@ -36,7 +36,5 @@ class LoginActivity: AppCompatActivity(), LifecycleOwner {
     }
 }
 
-fun String.isEmailValid(): Boolean {
-    return !TextUtils.isEmpty(this) && android.util.Patterns.EMAIL_ADDRESS.matcher(this).matches()
-}
+
